rename schema-validator folder to lib
