+++ lib/schema-validator-error.js
